HACK: Copy GbE region on gaze16-3060-b
gaze16-3060-b uses the onboard GbE controller instead of a Realtek
controller. The device stores important information, like the MAC
address, in a 8k region on the SPI flash at 0x1000-0x3000.

This is a just a quick fix. A more proper solution would be to read
FLREG3 from the IFD to verify where the GbE region is.

Signed-off-by: Tim Crawford <tcrawford@system76.com>